feat: Provide user facing types
